Add random PRNG seed for vacancies
Previous version had a hard-coded PRNG seed for populating the lattice with vacancies or defects. Default instead to seeding the main PRNG with a call to the hardware RNG. Setting the "lattice_rng_seed" parameter in the input file uses that value as the PRNG seed instead (if exact replicas are desired).