fix unconstrained Metropolis trial algorithm and unify call signatures
This fix restores the Markov property of the unconstrained Metropolis
trial algorithm. Before hitting the same spin twice in a single
iteration would restore the state of the spin to the state before the
first hit whether the first trial move has been accepted or not.
This breaks the Markov property of the trial moves, because the outcome
of the second trial move for a doubly hit spin depends on the outcome of
the first hit.

Additionally, this commit unifies the call signatures of the constrained
and unconstrained Metropolis algorithms.