calculate magnetization directly when initializing Monte Carlo MDC
The goal is to avoid stale cache data, because it is important for the
fixed magnetization direction to match the current magnetization
direction.
This approach cannot prevent the user from changing the direction of the
spins while the simulation is running. But that should be a scenario
that is expected to break simulations.