introduce alias `TableParserInit` and type trait `arrays_to_tuple`
The type trait `arrays_to_tuple` can convert a variadic pack of
std::array types into a flattened std::tuple of equivalent types.
This is useful for specifying repeating sequences of types in a variadic
parameter pack as are used in the `TableParser` class template.

The `TableParserInit` alias utilizes `arrays_to_tuple` to allow for a
significantly more compact and readable `TableParser` instantiation.